# src/main/java/co/edu/uniquindio/application/controllers/UsuarioControlador.java
package co.edu.uniquindio.application.controllers;


import co.edu.uniquindio.application.dtos.alojamiento.ItemAlojamientoDTO;
import co.edu.uniquindio.application.dtos.reserva.ItemReservaDTO;
import co.edu.uniquindio.application.dtos.usuario.CreacionAnfitrionDTO;
import co.edu.uniquindio.application.dtos.usuario.CreacionUsuarioDTO;
import co.edu.uniquindio.application.dtos.usuario.EdicionUsuarioDTO;
import co.edu.uniquindio.application.dtos.RespuestaDTO;
import co.edu.uniquindio.application.dtos.usuario.CambioContrasenaDTO;
import co.edu.uniquindio.application.dtos.usuario.UsuarioDTO;
import co.edu.uniquindio.application.services.AlojamientoServicio;
import co.edu.uniquindio.application.services.UsuarioServicio;
import jakarta.validation.Valid;
import lombok.RequiredArgsConstructor;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.*;

import java.util.List;


@RestController
@RequestMapping("/api/usuarios")
@RequiredArgsConstructor
public class UsuarioControlador {

    private final UsuarioServicio usuarioServicio;
    private final AlojamientoServicio alojamientoServicio;

    @PostMapping("/{id}/anfitrion")
    public ResponseEntity<RespuestaDTO<String>> crearAnfitrion(@PathVariable String id, @Valid @RequestBody CreacionAnfitrionDTO dto) throws Exception {
        usuarioServicio.crearAnfitrion(id, dto);
        return ResponseEntity.ok(new RespuestaDTO<>(false, "Usuario convertido a anfitrión exitosamente"));
    }

    @PutMapping("/{id}")
    public ResponseEntity<RespuestaDTO<String>> editar(@PathVariable String id, @Valid @RequestBody EdicionUsuarioDTO edicionUsuarioDTO) throws Exception {
        usuarioServicio.editar(id , edicionUsuarioDTO);
        return ResponseEntity.ok(new RespuestaDTO<>(false, "El usuario ha sido actualizado"));
    }

    @GetMapping("/{id}")
    public ResponseEntity<RespuestaDTO<UsuarioDTO>> obtener(@PathVariable String id) throws Exception {
        UsuarioDTO usuarioDTO = usuarioServicio.obtener(id);
        return ResponseEntity.ok(new RespuestaDTO<>(false, usuarioDTO));
    }

    @DeleteMapping("/{id}")
    public ResponseEntity<RespuestaDTO<String>> eliminar(@PathVariable String id) throws Exception {
        usuarioServicio.eliminar(id);
        return ResponseEntity.noContent().build();
    }

    @PatchMapping("/{id}/contrasena")
    public ResponseEntity<RespuestaDTO<String>> cambiarContrasena(@PathVariable String id, @Valid @RequestBody CambioContrasenaDTO dto) throws Exception {
        usuarioServicio.cambiarContrasena(id, dto);
        return ResponseEntity.ok(new RespuestaDTO<>(false, "La contraseña ha sido cambiada"));
    }

    @GetMapping("/{id}/alojamientos")
    public ResponseEntity<RespuestaDTO<List<ItemAlojamientoDTO>>> obtenerAlojamientosUsuario(@PathVariable String id) throws Exception {
        List<ItemAlojamientoDTO> alojamientos = alojamientoServicio.obtenerAlojamientoUsuario(id, 0);
        return ResponseEntity.ok(new RespuestaDTO<>(false, alojamientos));
    }

    @GetMapping("/{id}/reservas")
    public ResponseEntity<RespuestaDTO<List<ItemReservaDTO>>> obtenerReservasUsuario(@PathVariable String id) throws Exception {
        return ResponseEntity.ok(new RespuestaDTO(false, List.of()));
    }

}